+++ softmax_test.go
@@ -9,22 +9,22 @@ import (
 func TestSoftmaxLoss(t *testing.T) {
 	for i := 0; i < 2560; i++ {
 		size := i/10 + 2
-		vec := make([]float32, size)
-		target := make([]float32, len(vec))
+		vec := make([]float64, size)
+		target := make([]float64, len(vec))
 		for i := range vec {
-			vec[i] = float32(rand.NormFloat64())
-			target[i] = float32(rand.NormFloat64())
+			vec[i] = float64(rand.NormFloat64())
+			target[i] = float64(rand.NormFloat64())
 		}
 		if i%2 == 0 {
 			// Test extreme cases.
 			for j := range vec {
 				if rand.Intn(2) == 0 {
-					vec[j] *= float32(rand.NormFloat64() * 10)
+					vec[j] *= float64(rand.NormFloat64() * 10)
 				}
 			}
 		}
 		actual := SoftmaxLoss(vec, target)
-		expected := float32(0)
+		expected := float64(0)
 		log := logSoftmax(vec)
 		for i, l := range log {
 			expected -= l * target[i]
@@ -37,18 +37,18 @@ func TestSoftmaxLoss(t *testing.T) {
 
 func BenchmarkSoftmaxLoss(b *testing.B) {
 	b.Run("Size2", func(b *testing.B) {
-		params := []float32{-0.5, 0.7}
-		targets := []float32{0.2, 0.3}
+		params := []float64{-0.5, 0.7}
+		targets := []float64{0.2, 0.3}
 		for i := 0; i < b.N; i++ {
 			SoftmaxLoss(params, targets)
 		}
 	})
 	b.Run("Size10", func(b *testing.B) {
-		params := make([]float32, 10)
-		targets := make([]float32, 10)
+		params := make([]float64, 10)
+		targets := make([]float64, 10)
 		for i := range params {
-			params[i] = float32(rand.NormFloat64())
-			targets[i] = float32(rand.NormFloat64())
+			params[i] = float64(rand.NormFloat64())
+			targets[i] = float64(rand.NormFloat64())
 		}
 		for i := 0; i < b.N; i++ {
 			SoftmaxLoss(params, targets)
@@ -58,18 +58,18 @@ func BenchmarkSoftmaxLoss(b *testing.B) {
 
 func BenchmarkSoftmaxLossGrad(b *testing.B) {
 	for i := 0; i < b.N; i++ {
-		SoftmaxLossGrad([]float32{1.5, -0.3}, []float32{1.0, 0})
+		SoftmaxLossGrad([]float64{1.5, -0.3}, []float64{1.0, 0})
 	}
 }
 
 func BenchmarkSoftmaxLossDelta(b *testing.B) {
 	for i := 0; i < b.N; i++ {
-		SoftmaxLossDelta([]float32{1.5, -0.3}, []float32{1.0, 0}, []float32{0.5, -0.2}, 0.1)
+		SoftmaxLossDelta([]float64{1.5, -0.3}, []float64{1.0, 0}, []float64{0.5, -0.2}, 0.1)
 	}
 }
 
 func BenchmarkSoftmaxLossKL(b *testing.B) {
 	for i := 0; i < b.N; i++ {
-		SoftmaxLossKL([]float32{1.5, -0.3}, []float32{0.5, -0.2}, 0.1)
+		SoftmaxLossKL([]float64{1.5, -0.3}, []float64{0.5, -0.2}, 0.1)
 	}
 }
